Crew Refactors
