+++ app/modules/intelligence/agents/chat_agents/qna_agent.py
@@ -18,6 +18,7 @@ from sqlalchemy.orm import Session
 from app.modules.conversations.message.message_model import MessageType
 from app.modules.conversations.message.message_schema import NodeContext
 from app.modules.intelligence.agents.agents_service import AgentsService
+from app.modules.intelligence.agents.crew.rag_crew import kickoff_rag_crew
 from app.modules.intelligence.memory.chat_history_service import ChatHistoryService
 from app.modules.intelligence.prompts.classification_prompts import (
     AgentType,
@@ -27,9 +28,6 @@ from app.modules.intelligence.prompts.classification_prompts import (
 )
 from app.modules.intelligence.prompts.prompt_schema import PromptResponse, PromptType
 from app.modules.intelligence.prompts.prompt_service import PromptService
-from app.modules.intelligence.tools.agentic_tools.rag_agentic_tool import (
-    kickoff_rag_crew,
-)
 
 logger = logging.getLogger(__name__)
