+++ app/modules/intelligence/agents/chat_agents/unit_test_agent.py
@@ -17,6 +17,7 @@ from sqlalchemy.orm import Session
 from app.modules.conversations.message.message_model import MessageType
 from app.modules.conversations.message.message_schema import NodeContext
 from app.modules.intelligence.agents.agents_service import AgentsService
+from app.modules.intelligence.agents.crew.unit_test_crew import kickoff_unit_test_crew
 from app.modules.intelligence.memory.chat_history_service import ChatHistoryService
 from app.modules.intelligence.prompts.classification_prompts import (
     AgentType,
@@ -26,9 +27,6 @@ from app.modules.intelligence.prompts.classification_prompts import (
 )
 from app.modules.intelligence.prompts.prompt_schema import PromptResponse, PromptType
 from app.modules.intelligence.prompts.prompt_service import PromptService
-from app.modules.intelligence.tools.agentic_tools.unit_test_agentic_tool import (
-    kickoff_unit_test_crew,
-)
 from app.modules.intelligence.tools.kg_based_tools.get_code_from_node_id_tool import (
     GetCodeFromNodeIdTool,
 )
