+++ app/modules/intelligence/agents/crew/debug_rag_crew.py
@@ -14,6 +14,9 @@ from app.modules.intelligence.tools.kg_based_tools.get_code_from_multiple_node_i
     GetCodeFromMultipleNodeIdsTool,
     get_code_from_multiple_node_ids_tool,
 )
+from app.modules.intelligence.tools.kg_based_tools.get_code_from_node_id_tool import (
+    get_code_from_node_id_tool,
+)
 from app.modules.intelligence.tools.kg_based_tools.get_code_from_probable_node_name_tool import (
     get_code_from_probable_node_name_tool,
 )
@@ -35,11 +38,12 @@ class RAGResponse(BaseModel):
     response: List[NodeResponse]
 
 
-class RAGAgenticTool:
+class DebugRAGCrew:
     def __init__(self, sql_db, llm, mini_llm, user_id):
         self.openai_api_key = os.getenv("OPENAI_API_KEY")
         self.max_iter = os.getenv("MAX_ITER", 5)
         self.sql_db = sql_db
+        self.get_code_from_node_id = get_code_from_node_id_tool(sql_db, user_id)
         self.get_code_from_multiple_node_ids = get_code_from_multiple_node_ids_tool(
             sql_db, user_id
         )
@@ -169,7 +173,6 @@ class RAGAgenticTool:
                 "Markdown formatted chat response to user's query grounded in provided code context and tool results"
             ),
             agent=query_agent,
-            output_pydantic=RAGResponse,
             async_execution=True,
         )
 
@@ -216,7 +219,7 @@ class RAGAgenticTool:
         return result
 
 
-async def kickoff_rag_crew(
+async def kickoff_debug_rag_crew(
     query: str,
     project_id: str,
     chat_history: List,
@@ -226,9 +229,9 @@ async def kickoff_rag_crew(
     mini_llm,
     user_id: str,
 ) -> str:
-    rag_agent = RAGAgenticTool(sql_db, llm, mini_llm, user_id)
+    debug_agent = DebugRAGCrew(sql_db, llm, mini_llm, user_id)
     file_structure = GithubService(sql_db).get_project_structure(project_id)
-    result = await rag_agent.run(
+    result = await debug_agent.run(
         query, project_id, chat_history, node_ids, file_structure
     )
     return result
